+++ src/comm/drivers/LibUSB/DeviceHandleLibUSB.cpp
@@ -9,8 +9,8 @@
 
 namespace
 {
-uint16_t kLibUSBReadTimeout = 2;   // Timeout of a input bulk transfer  (0 = NO timeout)
-uint16_t kLibUSBWriteTimeout = 50; // Timeout of a output bulk transfer (0 = NO timeout)
+unsigned kLibUSBReadTimeout = 2;   // Timeout of a input bulk transfer  (0 = NO timeout)
+unsigned kLibUSBWriteTimeout = 50; // Timeout of a output bulk transfer (0 = NO timeout)
 } // namespace
 
 namespace sl
