+++ src/gfx/displays/GDisplayMaschineMikro.cpp
@@ -19,7 +19,7 @@ namespace cabl
 //--------------------------------------------------------------------------------------------------
 
 void GDisplayMaschineMikro::setPixel(
-  uint16_t x_, uint16_t y_, const util::ColorRGB& color_, bool bSetDirtyChunk_)
+  unsigned x_, unsigned y_, const util::ColorRGB& color_, bool bSetDirtyChunk_)
 {
   if (x_ >= width() || y_ >= height() || color_.transparent())
   {
@@ -53,7 +53,7 @@ void GDisplayMaschineMikro::setPixel(
 
 //--------------------------------------------------------------------------------------------------
 
-util::ColorRGB GDisplayMaschineMikro::pixel(uint16_t x_, uint16_t y_) const
+util::ColorRGB GDisplayMaschineMikro::pixel(unsigned x_, unsigned y_) const
 {
   if (x_ >= width() || y_ >= height())
   {
